\name{validCdfNames}
\alias{validCdfNames}

\title{ Supported annotation packages for crlmm genotyping}
\description{
Supported annotation packages for crlmm genotyping
}
\usage{
validCdfNames()
}
\details{
  List of available annotation packages
}
\value{
  character vector
}

\author{
R.Scharpf
}


\examples{
validCdfNames()
}
\keyword{manip}